[task-attribution] Move to an implicit GCed task container model
The current task_container_ is a circular buffer with limited space.
That limits the potential use cases for TaskAttribution, and adds
needless complexity. This CL significantly simplifies that model, by
moving the container to be implicit, and relying on GC to clean up tasks
once they are no longer needed - that is, once all their descendent
tasks were done and evicted.

Change-Id: I5d081a3d4208f3b942b9b7f77b539e56cf787c11